Refactor test files: fix naming, imports, patch counts, and remove duplicates
- Rename files with "and" in name or generic names to per-class/module files
- Move all imports to module level (no inside-test-method imports)
- Delete tests with more than 3 patches
- Remove tests that duplicate existing integration tests
- Split large multi-concern test files into focused single-class files
- Rename test_pipeline_infrastructure.py to test_data_pipeline_core.py
- Distribute tests from deleted files into appropriate existing files
