+++ tests/test_calibration_config.py
@@ -1,65 +1,13 @@
-"""Tests for CommandConfig and CalibrationConfig."""
-
-import tempfile
-from pathlib import Path
+"""Tests for CalibrationConfig, GradiometryConfig, and SetQualityAndNaNConfig."""
 
 import pytest
 import yaml
 
-from imap_mag.config.CalibrationConfig import CalibrationConfig, GradiometryConfig, SetQualityAndNaNConfig
-from imap_mag.config.CommandConfig import CommandConfig
-
-
-class TestCommandConfig:
-    def test_default_work_sub_folder_is_none(self):
-        config = CommandConfig()
-        assert config.work_sub_folder is None
-
-    def test_setup_work_folder_creates_directory_when_not_exists(self, tmp_path):
-        from unittest.mock import MagicMock
-
-        app_settings = MagicMock()
-        app_settings.work_folder = tmp_path / "work"
-
-        config = CommandConfig(work_sub_folder="mycommand")
-        result = config.setup_work_folder(app_settings)
-
-        assert result.exists()
-        assert result == tmp_path / "work" / "mycommand"
-
-    def test_setup_work_folder_uses_sub_folder_when_provided(self, tmp_path):
-        from unittest.mock import MagicMock
-
-        app_settings = MagicMock()
-        app_settings.work_folder = tmp_path
-
-        config = CommandConfig(work_sub_folder="subdir")
-        result = config.setup_work_folder(app_settings)
-
-        assert result == tmp_path / "subdir"
-
-    def test_setup_work_folder_returns_same_folder_on_second_call(self, tmp_path):
-        from unittest.mock import MagicMock
-
-        app_settings = MagicMock()
-        app_settings.work_folder = tmp_path
-
-        config = CommandConfig()
-        first = config.setup_work_folder(app_settings)
-        second = config.setup_work_folder(app_settings)
-
-        assert first == second
-
-    def test_setup_work_folder_without_sub_folder_uses_app_work_folder(self, tmp_path):
-        from unittest.mock import MagicMock
-
-        app_settings = MagicMock()
-        app_settings.work_folder = tmp_path
-
-        config = CommandConfig()
-        result = config.setup_work_folder(app_settings)
-
-        assert result == tmp_path
+from imap_mag.config.CalibrationConfig import (
+    CalibrationConfig,
+    GradiometryConfig,
+    SetQualityAndNaNConfig,
+)
 
 
 class TestGradiometryConfig:
@@ -80,7 +28,7 @@ class TestGradiometryConfig:
 class TestSetQualityAndNaNConfig:
     def test_csv_file_is_required(self):
         with pytest.raises(Exception):
-            SetQualityAndNaNConfig()  # no csv_file provided
+            SetQualityAndNaNConfig()
 
     def test_csv_file_is_stored(self):
         config = SetQualityAndNaNConfig(csv_file="quality.csv")
