+++ tests/test_data_pipeline_core.py
